changes to support decimals

+++ ecm1410-workshop-2-main/RectangleComputation.java
@@ -1,7 +1,7 @@
 public class RectangleComputation {
     public static void main(String[] args) {
-        double height = Integer.parseInt(args[0]);
-        double width = Integer.parseInt(args[1]);
+        double height = Double.parseDouble(args[0]);
+        double width = Double.parseDouble(args[1]);
         double area = height  * width;
         double perimeter = 2 * height + 2 * width;
         System.out.println("Shape: Rectangle");
